requirements.txt
anthropic
openai
python-dotenv
requests
markdown
tqdm
groq
